# server/openapi.yaml
openapi: "3.1.0"
info:
  version: 1.0.0
  title: Launchr
  description: An API of Launchr actions
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /customisation:
    get:
      summary: Customisation config
      operationId: getCustomisationConfig
      responses:
        '200':
          description: config response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customisation'
        default:
          $ref: '#/components/responses/DefaultError'
  /actions:
    get:
      summary: Lists all actions
      description: |
        Returns all launchr actions
      operationId: getActions
      responses:
        '200':
          description: actions response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionShort'
        default:
          $ref: '#/components/responses/DefaultError'
  /actions/{id}:
    get:
      summary: Returns action by id
      description: returns action by id
      operationId: getActionByID
      parameters:
        - $ref: '#/components/parameters/ActionId'
      responses:
        '200':
          description: action response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionFull'
        default:
          $ref: '#/components/responses/DefaultError'
    post:
      summary: runs action
      description: executes action
      operationId: runAction
      parameters:
        - $ref: '#/components/parameters/ActionId'
      requestBody:
        description: Action arguments and options
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionRunParams'
      responses:
        '201':
          description: action response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionRunInfo'
        default:
          $ref: '#/components/responses/DefaultError'
  /actions/{id}/schema.json:
    get:
      summary: Returns action json schema
      description: returns action json schema
      operationId: getActionJSONSchema
      parameters:
        - $ref: '#/components/parameters/ActionId'
      responses:
        '200':
          description: action json schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONSchema'
        default:
          $ref: '#/components/responses/DefaultError'
  /actions/{id}/running:
    get:
      summary: Returns running actions
      description: returns running actions
      operationId: getRunningActionsByID
      parameters:
        - $ref: '#/components/parameters/ActionId'
      responses:
        '200':
          description: action run info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionRunInfo'
        default:
          $ref: '#/components/responses/DefaultError'
  /actions/{id}/running/{runId}:
    get:
      summary: Returns action run info
      description: returns action run info
      operationId: getOneRunningActionByID
      parameters:
        - $ref: '#/components/parameters/ActionId'
        - $ref: '#/components/parameters/ActionRunInfoId'
      responses:
        '200':
          description: action run info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionRunInfo'
        default:
          $ref: '#/components/responses/DefaultError'
  /actions/{id}/running/{runId}/streams:
    get:
      summary: Returns running action streams
      description: returns running action streams
      operationId: getRunningActionStreams
      parameters:
        - $ref: '#/components/parameters/ActionId'
        - $ref: '#/components/parameters/ActionRunInfoId'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: action run info streams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActionRunStreamData'
        default:
          $ref: '#/components/responses/DefaultError'
components:
  responses:
    DefaultError:
      description: unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    ActionId:
      name: id
      in: path
      description: ID of action to fetch
      required: true
      schema:
        type: string
    ActionRunInfoId:
      name: runId
      in: path
      description: ID of the running action
      required: true
      schema:
        type: string
    Offset:
      name: offset
      in: query
      description: number of elements to skip
      schema:
        type: integer
    Limit:
      name: limit
      in: query
      description: number of elements to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
  schemas:
    Customisation:
      type: object
      x-go-name: "CustomisationConfig"
      x-go-type: "map[string]interface{}"
      x-go-type-skip-optional-pointer: true
    ActionShort:
      allOf:
        - type: object
          required:
            - id
            - title
            - description
          properties:
            id:
              type: string
              x-go-name: "ID"
            title:
              type: string
            description:
              type: string
    ActionFull:
      allOf:
        - $ref: '#/components/schemas/ActionShort'
        - type: object
          required:
            - jsonschema
          properties:
            jsonschema:
              type: object
              x-go-name: "JSONSchema"
              x-go-type: "jsonschema.Schema"
              x-go-type-import:
                path: "github.com/launchrctl/launchr/pkg/jsonschema"
            uischema:
              type: object
              x-go-name: "UISchema"
              x-go-type: "map[string]interface{}"
              x-go-type-skip-optional-pointer: true
    ActionRunParams:
      allOf:
        - type: object
          required:
            - arguments
            - options
          properties:
            arguments:
              type: object
              x-go-type: "action.TypeArgs"
              x-go-type-import:
                path: "github.com/launchrctl/launchr/pkg/action"
            options:
              type: object
              x-go-type: "action.TypeOpts"
              x-go-type-import:
                path: "github.com/launchrctl/launchr/pkg/action"
            changed:
              type: array
              items:
                type: string
    ActionRunInfo:
      allOf:
        - type: object
          required:
            - id
            - status
          properties:
            id:
              type: string
              x-go-name: "ID"
            status:
              $ref: '#/components/schemas/ActionRunStatus'
    ActionRunStatus:
      type: string
      enum:
        - created
        - running
        - finished
        - error
    ActionRunStreamData:
      allOf:
        - type: object
          required:
            - type
            - content
            - offset
            - count
          properties:
            type:
              type: string
              enum:
                - stdOut
                - stdIn
                - stdErr
            content:
              type: string
            offset:
              type: integer
            count:
              type: integer
    JSONSchema:
      type: object
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int
        message:
          type: string